♻️ Refactored get_vehicles function
